Add attribute deletion.

+++ app/code/community/FireGento/DeinstallMRG/controllers/Adminhtml/MrgDeinstallController.php
@@ -15,7 +15,8 @@ class FireGento_DeinstallMRG_Adminhtml_MrgDeinstallController
 
     public function deinstallAllAction()
     {
-        $this->_deinstallAll();
+        $this->_deinstallAllFiles();
+        $this->_deinstallAllDatabaseChanges();
         $this->_forward('index');
     }
 
@@ -43,8 +44,8 @@ class FireGento_DeinstallMRG_Adminhtml_MrgDeinstallController
     protected function _uninstallMageLocalWishlist($force)
     {
         if (
-            $this->getRequest()->getParam('mage_local_wishlist_abstract')
-            || $force
+            $force
+            || $this->getRequest()->getParam('mage_local_wishlist_abstract')
         ) {
             // TODO check path
             unlink(
@@ -65,7 +66,7 @@ class FireGento_DeinstallMRG_Adminhtml_MrgDeinstallController
         }
     }
 
-    protected function _deinstallAll()
+    protected function _deinstallAllFiles()
     {
         $directoriesToDelete = array(
             'app/code/community/Symmetrics/Agreement',
@@ -96,7 +97,7 @@ class FireGento_DeinstallMRG_Adminhtml_MrgDeinstallController
             'app/design/frontend/default/default/template/symmetrics/invoice',
             'app/design/frontend/default/default/template/tweaksgerman',
             'app/etc/modules/Symmetrics_Agreement.xml',
-            'app/etc/modules/Symmetrics_ConfigGerman.xml',
+            'app/etc/modules/Symmetrics_Config.xml',
             'app/etc/modules/Symmetrics_ConfigGermanTexts.xml',
             'app/etc/modules/Symmetrics_DeliveryTime.xml',
             'app/etc/modules/Symmetrics_Imprint.xml',
@@ -111,7 +112,6 @@ class FireGento_DeinstallMRG_Adminhtml_MrgDeinstallController
             'js/symmetrics/stockindicator',
             'js/symmetrics/tweaksgerman',
             'lib/Symmetrics/dompdf/',
-            'skin/adminhtml/default/default/images/symmetrics/aws',
             'skin/adminhtml/default/default/images/symmetrics/awsprovider.png',
             'skin/adminhtml/default/default/images/symmetrics/bottomhr.png',
             'skin/adminhtml/default/default/images/symmetrics/boxbg.png',
@@ -138,13 +138,26 @@ class FireGento_DeinstallMRG_Adminhtml_MrgDeinstallController
     }
 
 
+    /**
+     * Delete a directory or file recursive
+     *
+     * @param $dir
+     */
     private function delTree($dir)
     {
+        if (!file_exists($dir)) {
+            // if the path is already missing we can skip it
+            return;
+        }
+
         if (is_file($dir)) {
+            // if it is a file, we just delete it
             unlink($dir);
             return;
         }
 
+        // if it is a directory, we loop through it an call this
+        // method recursive
         foreach (new DirectoryIterator($dir) as $directory) {
             /* @var $directory DirectoryIterator */
             if (!$directory->isDot()) {
@@ -152,7 +165,47 @@ class FireGento_DeinstallMRG_Adminhtml_MrgDeinstallController
             }
         }
 
+        // when we deleted all directory and files, we can
+        // delete the directory itself
         rmdir($dir);
+    }
+
+    protected function _deinstallAllDatabaseChanges()
+    {
+        /* @var $installer Mage_Eav_Model_Entity_Setup */
+        $installer = Mage::getResourceModel(
+            'eav/entity_setup', 'core_setup'
+        );
+
+//app\code\community\Symmetrics\ConfigGerman\sql\config_german_setup\mysql4-install-0.1.0.php
+//TODO — remove the wight attribute out of catalog_product?
+//
+//app\code\community\Symmetrics\ConfigGermanTexts\sql\config_german_texts_setup\mysql4-install-0.1.0.php
+//TODO — $this->updateFooterLinksBlock($data); => Keine Ahnung was das ist, Andi fragen
+//
+//app\code\community\Symmetrics\DeliveryTime\sql\deliverytime_setup\mysql4-install-0.2.1.php
+//TODO — remove delivery_time product attribute from catalog_product?
+//
+//    app\code\community\Symmetrics\Imprint\sql\imprint_setup\mysql4-install-0.2.0.php
+// TODO — Changes a lot of configuration data – needs to be reviewed!
+//
+//    app\code\community\Symmetrics\PdfPrinter\sql\pdfprinter_setup\mysql4-install-0.1.0.php
+// TODO — Remove Mage::getBaseDir('media') . DS . 'pdfprinter'
+//
+//app\code\community\Symmetrics\SecurePassword\sql\securepassword_setup\mysql4-install-0.1.0.php
+
+        $installer->startSetup();
+        $installer->removeAttribute('customer', 'failed_logins');
+        $installer->removeAttribute('customer', 'last_failed_login');
+        $installer->removeAttribute('customer', 'last_unlock_time');
+        $installer->removeAttribute('customer', 'unlock_customer');
+
+
+//app\code\community\Symmetrics\SetMeta\sql\setmeta_setup\mysql4-install-0.2.0.php
+        $installer->removeAttribute('catalog_product', 'generate_meta');
+        $installer->endSetup();
+
+
     }
 
 }
